writecache: have some sleep between iterations, fix #3642
Once an object got into the cache c.objCounters.Size() is not zero until it
leaves the cache, even if there are no new objects coming in. This means we're
looping here without any stops while we have objects in cache and a lot of
cycles can be made even with a very small ops/s load (multiplied by the number
of shards). This loop also makes a copy of object list on each iteration which
leads to _a lot_ of allocations and associated GC pressure.

As a minimal fix we can just iterate over the list once per 100ms, 20 threads
can add enough pressure to real HDDs for them to be busy all this time and we
won't be wasting cycles for nothing.

Signed-off-by: Roman Khimov <roman@nspcc.ru>