sn/object: Do not add session token to EC part
The session token is in the parent object, but in the part it is
meaningless. Moreover, if the part object needs to be recreated, the
session token will no longer exist. Therefore, not adding a token
preserves the object ID which is good for deduplication.

For this to work, it was necessary to make an exception when checking
the signature against the object owner.

Refs #3628.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>