pkg/timers: do not skip delta handlers ever
It was intended to prevent apps from making useless handler calls when they are
outdated. Still, in practice, it introduced a race condition and skips events
that must occur immediately at the new epoch event or near it. Treat every
sub-epoch timer as not-called-yet at every `Reset` call; calling at the wrong
time (when there is an incredible latency) is less error-prone than not calling
at all. Closes #3652.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>