ir/container: Implement iterative background structuring in contract
Within https://github.com/nspcc-dev/neofs-contract/issues/449, there is
a need to store containers as VM structs. Since
https://github.com/nspcc-dev/neofs-contract/pull/534, struct items are
stored for new containers. This also needs to be done for containers
created before the upgrade.

Since doing this in a contract update transaction turned out to be too
GAS-intensive, this implements structuring through the IR background
process. Specialized method is called iteratively doing the same thing,
but for several containers at a time.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>