writecache: fix potential double flush error
The following errors were detected during performance testing:

    warn	writecache/flush.go:50	flush scheduler paused due to error	{"shard_id": "DVLS7az3U6AY5qZ8GHwsaP", "substorage": "write-cache", "delay": 10}
    error	writecache/flush.go:167	can't flush objects	{"shard_id": "DVLS7az3U6AY5qZ8GHwsaP", "substorage": "write-cache", "worker": 6, "first_object": "8dsJKm3xFpiDS8etCjaYaEBxPdgmYZJJtfjvuWFMHmzv/E5k23T1EZuyzjY67p3PvzbJmGfL6XxkHjxBiE536244R", "error": "logical error: status: code = 2049 message = object not found"}

This could potentially happen because of double-deletion of addrs from
flushObjs, flushScheduler() can fetch the same addrs twice because of that.
An easy fix is to nullify addrs in post-flush case, but we better simplify
lusher thread logic and avoid duplication at all.

Signed-off-by: Roman Khimov <roman@nspcc.ru>