engine: fix flaky TestSplitObjectExpirationWithLinkNotFound
For the scenario to happen properly we need LINK object to be present
in metabase, but missing from real FSTree. Only this way we get a LINK
in SplitInfo, but then fail to fetch it. Tombstones or garbage marks
leave physical LINK in place and all of its metadata as well, while
dropping with force removes both physical object and its meta, so
the scenario becomes the same as in TestSplitObjectExpirationWithoutLink.

Fixes #3669.

Signed-off-by: Roman Khimov <roman@nspcc.ru>