engine: collect child objects without search
In case LINK is missing we're using search to find objects that need to be
deleted, but this can work incorrectly for expired children since search
never returns outdated objects. I believe this fixes #3715, that's the
scenario (and log) that can be seen in UTs and it leads to the last object
part being stuck in the GC loop: parent object should be deleted, but
searching doesn't return an outdated object.

The new API is also somewhat more efficient.

Signed-off-by: Roman Khimov <roman@nspcc.ru>