metabase: parent status can be more appropriate for exists()
Inhume() sets both graveyard and garbage marks for deleted objects and when
checking for status graveyard has a precedence, so "tombstoned" is returned
for an object instead of "gcmarked". Now if graveyard marks are omitted we
still find a tombstone for simple objects, but childs only have a GC mark,
so unless we're to check for their parent we can't return better status.

Signed-off-by: Roman Khimov <roman@nspcc.ru>