metabase: drop Inhume()
It's not used in production code now. This makes some tests totally obsolete,
so they're removed. There is also some disagreement over logical counter
meaning in the code (most of Inhume() calls are replaced with TS Put()), so
for now we make it more consistent by counting Regular objects only.

Signed-off-by: Roman Khimov <roman@nspcc.ru>