storage/metabase: Return split error from `ResolveECPartWithPayloadLen()`
Previously, method returned 404 error if size-split object was requested.
Although this is actually a failure (size-split roots are not
erasure-coded), this error caused handling all shards in
`StorageEngine.GetECPartRange()` instead of abort.

This makes metabase to return split error in described case, and storage
engine to abort on it. As bonus, the method is in sync with
`ResolveECPart()` now.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>